requirements.txt
customtkinter
python-docx